requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
lxml==5.1.0
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
undetected-chromedriver==3.5.4
urllib3==2.2.0
websockets==12.0
wsproto==1.2.0
